+++ lib/src/clojure_lsp/queries.clj
@@ -98,6 +98,7 @@
 (def xf-analysis->protocol-impls (xf-analysis->bucket-elems :protocol-impls))
 (def xf-analysis->var-definitions (xf-analysis->bucket-elems :var-definitions))
 (def xf-analysis->var-usages (xf-analysis->bucket-elems :var-usages))
+(def xf-analysis->var-usages-and-symbols (xf-analysis->buckets-elems :var-usages :symbols))
 (def xf-analysis->symbols (xf-analysis->bucket-elems :symbols))
 (def xf-analysis->vars (xf-analysis->buckets-elems :var-definitions :var-usages :symbols))
 
@@ -518,7 +519,7 @@
   (let [names (var-definition-names var-definition)]
     (into []
           (comp
-            (if include-declaration? xf-analysis->vars xf-analysis->var-usages)
+            (if include-declaration? xf-analysis->vars xf-analysis->var-usages-and-symbols)
             (filter #(contains? names (:name %)))
             (filter #(safe-equal? (:ns var-definition) (or (:ns %) (:to %))))
             (filter #(or include-declaration?
@@ -566,8 +567,8 @@
                    (internal-analysis db))
         lang (:lang quoted-symbol)
         lang (if (= :edn lang)
-                     ;; when referring to qualified-symbols in edn, pretend it's
-                     ;; referenced from JVM Clojure
+               ;; when referring to qualified-symbols in edn, pretend it's
+               ;; referenced from JVM Clojure
                :clj
                lang)
         name (:name quoted-symbol)
@@ -686,27 +687,6 @@
           (xf-analysis->scoped-path-analysis path-uri))
         (:analysis db)))
 
-(def ^:private xf-defmethods
-  (comp (filter :defmethod)
-        (medley/distinct-by (juxt :to :name :name-row :name-col))))
-
-(defn find-defmethods [db uri]
-  (into []
-        xf-defmethods
-        (get-in db [:analysis uri :var-usages])))
-
-(defn find-internal-definitions
-  "All ns definitions, var definitions and defmethods."
-  [db]
-  (let [analysis (internal-analysis db)]
-    (concat (into []
-                  (xf-analysis->buckets-elems :namespace-definitions :var-definitions)
-                  analysis)
-            (into []
-                  (comp xf-analysis->var-usages
-                        xf-defmethods)
-                  analysis))))
-
 (defn find-keyword-definitions [db uri]
   (into []
         (medley/distinct-by (juxt :ns :name :row :col))
@@ -718,6 +698,35 @@
           xf-analysis->keyword-definitions
           (medley/distinct-by (juxt :ns :name :row :col)))
         (:analysis db)))
+
+(def ^:private xf-defmethods
+  (comp (filter :defmethod)
+        (medley/distinct-by (juxt :to :name :name-row :name-col))))
+
+(defn find-defmethods [db uri]
+  (into []
+        xf-defmethods
+        (get-in db [:analysis uri :var-usages])))
+
+(defn find-element-definitions
+  [db uri]
+  (concat []
+          (some-> (find-namespace-definition-by-uri db uri) vector)
+          (find-var-definitions db uri true)
+          (find-keyword-definitions db uri)
+          (find-defmethods db uri)))
+
+(defn find-all-element-definitions
+  "All ns definitions, var definitions and defmethods."
+  [db]
+  (let [analysis (internal-analysis db)]
+    (concat (into []
+                  (xf-analysis->buckets-elems :namespace-definitions :var-definitions)
+                  analysis)
+            (into []
+                  (comp xf-analysis->var-usages
+                        xf-defmethods)
+                  analysis))))
 
 (defn find-local-by-destructured-keyword [db uri keyword-element]
   (find-first (filter #(and (= (:name-row %) (:name-row keyword-element))
@@ -900,5 +909,6 @@
     :var-usages (if (:defmethod el)
                   :function
                   :variable)
+    :keyword-definitions :function
     :keyword-usages :field
     :null))