+++ lib/test/clojure_lsp/feature/paredit_test.clj
@@ -5,11 +5,14 @@
    [clojure.test :refer [deftest is]]))
 
 (defmacro ^:private assert-op [op expected code]
-  `(let [applied# (~op (:zloc (h/load-code-into-zloc-and-position ~code)))
+  `(let [applied# (~op h/default-uri (:zloc (h/load-code-into-zloc-and-position ~code)))
          [text# _#] (h/positions-from-text ~expected)]
-     (is (= text# (h/changes->code applied# (h/db))))))
+     (is (= text# (h/changes->code (-> applied# :changes-by-uri first second) (h/db))))))
 
 (deftest forward-slurp-test
+  (assert-op f.paredit/forward-slurp
+             (h/code "[1 [2 3| 4]]")
+             (h/code "[1 [2 3|] 4]"))
   (assert-op f.paredit/forward-slurp
              (h/code "[1 [2 [|3 4 5]] 6]")
              (h/code "[1 [2 [|3 4] 5] 6]"))
@@ -21,9 +24,18 @@
              (h/code "|1"))
   (assert-op f.paredit/forward-slurp
              (h/code "|")
-             (h/code "|")))
+             (h/code "|"))
+  (assert-op f.paredit/forward-slurp
+             (h/code "{:a {|:b 1} 2}")
+             (h/code "{:a {|:b 1}} 2"))
+  (assert-op f.paredit/forward-slurp
+             (h/code "{:a {:b {|:c :d}} 2}")
+             (h/code "{:a {:b {|:c :d}}} 2")))
 
 (deftest forward-barf-test
+  (assert-op f.paredit/forward-barf
+             (h/code "[1 [2] 3| 4]")
+             (h/code "[1 [2 3|] 4]"))
   (assert-op f.paredit/forward-barf
              (h/code "[1 [2 [|3 4] 5] 6]")
              (h/code "[1 [2 [|3 4 5]] 6]"))
@@ -39,6 +51,9 @@
 
 (deftest backward-slurp-test
   (assert-op f.paredit/backward-slurp
+             (h/code "[[1 |2 3] 4]")
+             (h/code "[1 [|2 3] 4]"))
+  (assert-op f.paredit/backward-slurp
              (h/code "[1 [[2 |3 4] 5] 6]")
              (h/code "[1 [2 [|3 4] 5] 6]"))
   (assert-op f.paredit/backward-slurp
@@ -51,7 +66,12 @@
              (h/code "|")
              (h/code "|")))
 
+[[1 2 3] 4]
+
 (deftest backward-barf-test
+  (assert-op f.paredit/backward-barf
+             (h/code "[1 |2 [3] 4]")
+             (h/code "[1 [|2 3] 4]"))
   (assert-op f.paredit/backward-barf
              (h/code "[1 [2 [|3 4] 5] 6]")
              (h/code "[1 [[2 |3 4] 5] 6]"))
@@ -81,7 +101,7 @@
              (h/code "[1 [2 [|] 5] 6]")
              (h/code "[1 [2 [|3 4] 5] 6]"))
   (assert-op f.paredit/kill
-             (h/code "|")
+             (h/code "|1")
              (h/code "|1"))
   (assert-op f.paredit/kill
              (h/code "|")